Move LintTest interface to a separate module
