+++ detectors/build.gradle
@@ -23,8 +23,9 @@ dependencies {
     compileOnly 'com.android.tools.lint:lint-api:26.3.0'
     compileOnly 'com.android.tools.lint:lint-checks:26.3.0'
 
-    testCompile 'junit:junit:4.12'
     testCompile project(":dsl")
+    testCompile project(":test")
+    testCompile 'junit:junit:4.12'
     testCompile 'org.assertj:assertj-core:3.3.0'
     testCompile 'org.mockito:mockito-core:2.23.0'
     testCompile 'com.android.tools.lint:lint:26.3.0'
